Add divs for flex containers and css styling
I really should've committed a lot more than this, but I just got so
carried away. The header and the footer are pushed all the way to the
top and end, but when I zoom out on the page there are large gaps above
and below them. Ideally, I would like them to stick to their respected
areas no matter how far I zoom out, as in thecss-exercises/foundations/
flex/06-flex-layout assignment. I may come back and fix this later.